feat: use LeanIMT, rather than BinaryIMT.
This commit replaces BinaryIMT with LeanIMT, which resolved the errors in the previous commit.

I've also removed the inputs array from the E3 struct, so inputs are no longer stored on-chain.

There seems to be some limit to how large items in the trees can be. A hack solution is to divide the inputHash by `10 ** 39`.