fix: deploy Poseidon
In previous commits, I was hitting errors with the Poseidon library. It seems to be an issue with the deterministic deployments and hardhat network, being tracked [here](https://github.com/privacy-scaling-explorations/zk-kit.solidity/issues/38). By deploying the Poseidon Lib directly, the issue was resolved. Note that this required adding an override to the compiler settings.